+++ core/src/storage/s3/S3ClientWrapper.h
@@ -63,11 +63,11 @@ class S3ClientWrapper : public IStorage {
     std::shared_ptr<Aws::S3::S3Client> client_ptr_;
     Aws::SDKOptions options_;
 
-    std::string minio_address_;
-    std::string minio_port_;
-    std::string minio_access_key_;
-    std::string minio_secret_key_;
-    std::string minio_bucket_;
+    std::string s3_address_;
+    std::string s3_port_;
+    std::string s3_access_key_;
+    std::string s3_secret_key_;
+    std::string s3_bucket_;
 };
 
 }  // namespace storage